Feature/301021 yan lee requestlogic changedepartment addnewmanager

+++ src/main/java/gui/GuiAdmin.java
@@ -1,7 +1,6 @@
 package gui;
 
-import dao.AccountDao;
-import dao.EmployeeDao;
+import dao.*;
 import entity.Account;
 import entity.Role;
 import gui.Form.GuiCreateAccountForm;
@@ -61,6 +60,8 @@ public class GuiAdmin extends JFrame {
 
         employeeName.setText(acc.getUserName());
         employeeRole.setText(role.getRoleTitle());
+        tableShowAccount.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
+        tableShowEmployee.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
 
         btn1.setCursor(new Cursor(Cursor.HAND_CURSOR));
         btn1.addMouseListener(new MouseAdapter() {
@@ -118,6 +119,16 @@ public class GuiAdmin extends JFrame {
                 btnDeleteActionPerformed(e);
             }
         });
+        contentPane.addMouseListener(new MouseAdapter() {
+            public void mouseClicked(MouseEvent e) {
+                ContentPaneMouseClicked(e);
+            }
+        });
+    }
+//deselect tables
+    private void ContentPaneMouseClicked(MouseEvent e) {
+        tableShowAccount.clearSelection();
+        tableShowEmployee.clearSelection();
     }
 
 
@@ -173,26 +184,61 @@ public class GuiAdmin extends JFrame {
 
     private void btnSearchActionPerformed(ActionEvent e) {
         if (txtSearchByName.getText().isEmpty()) {
-            JOptionPane.showMessageDialog(null, "Please input User Name");
-            tableShowAccount.repaint();
-        } else {
-            try {
-                Account acc = Account.getAccountFromUserName(txtSearchByName.getText().trim().replaceAll(("\\s"), ""));
-                String[] columnTitle = {"EMPLOYEE ID", "ROLE ID", "USER NAME", "PASSWORD"};
-                DefaultTableModel model = new DefaultTableModel(columnTitle, 0);
-
-                model.addRow(new Object[]{
-                        acc.getEmloyeeId(),
-                        acc.getRoleId(),
-                        acc.getUserName(),
-                        acc.getPassword()
-                });
-
-                tableShowAccount.setModel(model);
-            } catch (Exception em) {
-                JOptionPane.showMessageDialog(null, "Account does not exists!");
-                txtSearchByName.setText("");
+//            JOptionPane.showMessageDialog(null, "Please input User Name");
+//            tableShowAccount.repaint();
+            if(tabbedPaneShow.getSelectedIndex() == 0){
+                showListAccount();
+            }else{
+                showListEmployee();
             }
+        } else {
+            if(tabbedPaneShow.getSelectedIndex() == 0){
+                var accountDao = new AccountDao();
+                try {
+                    var listAcc = accountDao.getSearchedUserName(txtSearchByName.getText().trim().replaceAll(("\\s"), ""));
+                    String[] columnTitle = {"EMPLOYEE ID", "ROLE ID", "USER NAME", "PASSWORD"};
+                    DefaultTableModel model = new DefaultTableModel(columnTitle, 0);
+
+                    listAcc.forEach(user -> model.addRow(new Object[]{
+                            user.getEmloyeeId(),
+                            user.getRoleId(),
+                            user.getUserName(),
+                            user.getPassword()
+                    }));
+
+                    tableShowAccount.setModel(model);
+                } catch (Exception em) {
+                    JOptionPane.showMessageDialog(null, "Account does not exists!");
+                    txtSearchByName.setText("");
+                }
+            }else{
+                var employeeDao = new EmployeeDao();
+                try {
+                    var listEmployee = employeeDao.getSearchedEmployee(txtSearchByName.getText().trim().replaceAll(("\\s"), ""));
+                    String[] columnTitle = {"EMPLOYEE ID", "DEPARTMENT ID", "FULL NAME", "GENDER", "DATE OF BIRTH", "PHONE",
+                            "EMAIL", "DATE START", "ANNUAL LEAVE", "MANAGER ID"};
+                    DefaultTableModel model = new DefaultTableModel(columnTitle, 0);
+
+                    listEmployee.forEach(employee -> model.addRow(new Object[]{
+                            employee.getEmployeeId(),
+                            employee.getDepartmentId(),
+                            employee.getFullName(),
+                            employee.getGender(),
+                            employee.getDateOfBirth(),
+                            employee.getPhone(),
+                            employee.getEmail(),
+                            employee.getDateStart(),
+                            employee.getAnnualLeave(),
+                            employee.getManagerId()
+                    }));
+
+                    tableShowEmployee.setModel(model);
+                } catch (Exception em) {
+                    JOptionPane.showMessageDialog(null, "Account does not exists!");
+                    txtSearchByName.setText("");
+                }
+            }
+
         }
     }
 
@@ -263,17 +309,49 @@ public class GuiAdmin extends JFrame {
     }
 
     private void btnDeleteActionPerformed(MouseEvent e) {
-        int result = JOptionPane.showOptionDialog(null, "Are you sure exit?", "Exit", JOptionPane.YES_NO_OPTION,
+        int result = JOptionPane.showOptionDialog(null, "Are you sure you want to delete?", "Delete", JOptionPane.YES_NO_OPTION,
                 JOptionPane.PLAIN_MESSAGE, null, null, null);
 
         if (result == JOptionPane.YES_OPTION) {
             var dao = new AccountDao();
-            var rowindex = tableShowAccount.getSelectedRow();
-            var employeeId = tableShowAccount.getValueAt(rowindex, 0).toString();
-
             var account = new Account();
-            account.setEmloyeeId(Integer.parseInt(employeeId));
-            dao.deleteAccount(account);
+            var requestLeaveDao = new RequestLeaveDao();
+            var annualLeaveDao = new AnnualLeaveDao();
+            var departmentDao = new DepartmentDao();
+            var employeeDao = new EmployeeDao();
+
+            if(tableShowAccount.getSelectedRow() != -1){
+
+                var rowindex = tableShowAccount.getSelectedRow();
+                var employeeId = tableShowAccount.getValueAt(rowindex, 0).toString();
+                if(dao.getAdminID() == Integer.parseInt(employeeId)){
+                    JOptionPane.showMessageDialog(null, "you cant delete admin's account.");
+                }else{
+                    account.setEmloyeeId(Integer.parseInt(employeeId));
+                    dao.deleteAccount(account);
+                }
+                showListAccount();
+            }else if(tableShowEmployee.getSelectedRow() != -1){
+                var rowindex = tableShowEmployee.getSelectedRow();
+                var employeeID = tableShowEmployee.getValueAt(rowindex, 0).toString();
+
+                if(dao.getAdminID() == Integer.parseInt(employeeID)){
+                    JOptionPane.showMessageDialog(null, "you cant delete admin's employee info.");
+                }else if(!departmentDao.checkIfIsChiefDepartment(Integer.parseInt(employeeID))){
+                    requestLeaveDao.deleteRequestLeaveByEmployeeID(Integer.parseInt(employeeID));
+                    annualLeaveDao.deleteAnnualLeaveByEmployeeID(Integer.parseInt(employeeID));
+                    account.setEmloyeeId(Integer.parseInt(employeeID));
+                    dao.deleteAccount(account);
+                    employeeDao.deleteEmployeeByEmployeeID(Integer.parseInt(employeeID));
+                    showListEmployee();
+                    showListAccount();
+                }else{
+                    JOptionPane.showMessageDialog(null, "you cant delete a chief of department.");
+                }
+
+            }else {
+                JOptionPane.showMessageDialog(null, "Please select row you want to delete.");
+            }
         }
     }
